test: Presenter 테스트 대역 파일 분리
Presenter 테스트 파일에는 시나리오와 검증만 남기고 Repository 대역과 테스트 fixture를 역할별 파일로 분리했다.\n\n각 feature의 대역은 test source set 내부에서만 사용하도록 유지해 테스트 의도를 읽기 쉽게 정리했다.
