+++ feature/onboarding/src/test/kotlin/com/nexters/bandalart/feature/onboarding/presenter/OnboardingPresenterTest.kt
@@ -1,12 +1,9 @@
 package com.nexters.bandalart.feature.onboarding.presenter
 
-import com.nexters.bandalart.core.domain.repository.OnboardingRepository
 import com.nexters.bandalart.feature.home.HomeScreen
 import com.nexters.bandalart.feature.onboarding.OnboardingScreen
 import com.slack.circuit.test.FakeNavigator
 import com.slack.circuit.test.test
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
@@ -26,15 +23,4 @@ class OnboardingPresenterTest {
             assertTrue(repository.isCompleted)
         }
     }
-}
-
-private class FakeOnboardingRepository : OnboardingRepository {
-    var isCompleted = false
-        private set
-
-    override suspend fun setOnboardingCompletedStatus(flag: Boolean) {
-        isCompleted = flag
-    }
-
-    override fun flowIsOnboardingCompleted(): Flow<Boolean> = flowOf(isCompleted)
 }
