+++ feature/splash/src/test/kotlin/com/nexters/bandalart/feature/splash/presenter/SplashPresenterTest.kt
@@ -1,14 +1,11 @@
 package com.nexters.bandalart.feature.splash.presenter
 
-import com.nexters.bandalart.core.domain.repository.OnboardingRepository
 import com.nexters.bandalart.feature.home.HomeScreen
 import com.nexters.bandalart.feature.onboarding.OnboardingScreen
 import com.nexters.bandalart.feature.splash.SplashScreen
 import com.slack.circuit.runtime.screen.Screen
 import com.slack.circuit.test.FakeNavigator
 import com.slack.circuit.test.test
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertEquals
 import org.junit.Test
@@ -40,14 +37,4 @@ class SplashPresenterTest {
             assertEquals(expected, navigator.awaitResetRoot().newRoot)
         }
     }
-}
-
-private class FakeOnboardingRepository(isCompleted: Boolean) : OnboardingRepository {
-    private val completed = MutableStateFlow(isCompleted)
-
-    override suspend fun setOnboardingCompletedStatus(flag: Boolean) {
-        completed.value = flag
-    }
-
-    override fun flowIsOnboardingCompleted(): Flow<Boolean> = completed
 }
